refactor: 불필요한 파일 삭제 및 Dao 인터페이스 추가
