+++ src/main/java/wooteco/subway/dao/StationDaoImpl.java
@@ -1,0 +1,77 @@
+package wooteco.subway.dao;
+
+import java.util.List;
+import java.util.Optional;
+import javax.sql.DataSource;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import wooteco.subway.domain.Station;
+
+@Repository
+public class StationDaoImpl implements StationDao {
+
+    private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
+    private final SimpleJdbcInsert simpleJdbcInsert;
+    private final RowMapper<Station> stationRowMapper = (resultSet, rowNum) ->
+        new Station(
+            resultSet.getLong("id"),
+            resultSet.getString("name")
+        );
+
+    public StationDaoImpl(final DataSource dataSource) {
+        this.namedParameterJdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
+        this.simpleJdbcInsert = new SimpleJdbcInsert(dataSource)
+            .withTableName("station")
+            .usingGeneratedKeyColumns("id");
+    }
+
+    @Override
+    public Station save(final String name) {
+        checkDuplicateName(name);
+        SqlParameterSource parameters = new MapSqlParameterSource("name", name);
+        Long id = simpleJdbcInsert.executeAndReturnKey(parameters).longValue();
+        return new Station(id, name);
+    }
+
+    private void checkDuplicateName(String name) {
+        if (findByName(name).isPresent()) {
+            throw new IllegalArgumentException("이미 존재하는 지하철역 이름입니다.");
+        }
+    }
+
+    private Optional<Station> findByName(String name) {
+        String sql = "SELECT * FROM station WHERE name = :name";
+        MapSqlParameterSource parameters = new MapSqlParameterSource("name", name);
+        try {
+            return Optional.ofNullable(namedParameterJdbcTemplate.queryForObject(sql, parameters, stationRowMapper));
+        } catch (EmptyResultDataAccessException e) {
+            return Optional.empty();
+        }
+    }
+
+    @Override
+    public void deleteAll() {
+        String sql = "TRUNCATE TABLE station";
+        namedParameterJdbcTemplate.update(sql, new MapSqlParameterSource());
+        String resetIdSql = "ALTER TABLE station ALTER COLUMN id RESTART WITH 1";
+        namedParameterJdbcTemplate.update(resetIdSql, new MapSqlParameterSource());
+    }
+
+    @Override
+    public List<Station> findAll() {
+        String sql = "SELECT * FROM station";
+        return namedParameterJdbcTemplate.query(sql, stationRowMapper);
+    }
+
+    @Override
+    public void deleteById(Long id) {
+        String sql = "DELETE FROM station WHERE id = :id";
+        MapSqlParameterSource parameters = new MapSqlParameterSource("id", id);
+        namedParameterJdbcTemplate.update(sql, parameters);
+    }
+}
