+++ src/main/java/wooteco/subway/service/LineService.java
@@ -2,10 +2,8 @@ package wooteco.subway.service;
 
 import java.util.List;
 import java.util.stream.Collectors;
-
 import org.springframework.stereotype.Service;
-
-import wooteco.subway.dao.LineDao;
+import wooteco.subway.dao.LineDaoImpl;
 import wooteco.subway.domain.Line;
 import wooteco.subway.dto.LineRequest;
 import wooteco.subway.dto.LineResponse;
@@ -13,35 +11,35 @@ import wooteco.subway.dto.LineResponse;
 @Service
 public class LineService {
 
-    private final LineDao lineDao;
+    private final LineDaoImpl lineDaoImpl;
 
-    public LineService(LineDao lineDao) {
-        this.lineDao = lineDao;
+    public LineService(LineDaoImpl lineDaoImpl) {
+        this.lineDaoImpl = lineDaoImpl;
     }
 
     public LineResponse create(LineRequest lineRequest) {
-        Line line = lineDao.save(lineRequest.getName(), lineRequest.getColor());
+        Line line = lineDaoImpl.save(lineRequest.getName(), lineRequest.getColor());
         return new LineResponse(line);
     }
 
     public List<LineResponse> findAll() {
-        List<Line> lines = lineDao.findAll();
+        List<Line> lines = lineDaoImpl.findAll();
         return lines.stream()
-                .map(LineResponse::new)
-                .collect(Collectors.toList());
+            .map(LineResponse::new)
+            .collect(Collectors.toList());
     }
 
     public LineResponse findById(Long id) {
-        Line line = lineDao.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("해당 노선이 없습니다."));
+        Line line = lineDaoImpl.findById(id)
+            .orElseThrow(() -> new IllegalArgumentException("해당 노선이 없습니다."));
         return new LineResponse(line);
     }
 
     public void update(Long id, LineRequest lineRequest) {
-        lineDao.update(new Line(id, lineRequest.getName(), lineRequest.getColor()));
+        lineDaoImpl.update(new Line(id, lineRequest.getName(), lineRequest.getColor()));
     }
 
     public void delete(Long id) {
-        lineDao.deleteById(id);
+        lineDaoImpl.deleteById(id);
     }
 }
