+++ src/main/java/wooteco/subway/service/StationService.java
@@ -2,10 +2,8 @@ package wooteco.subway.service;
 
 import java.util.List;
 import java.util.stream.Collectors;
-
 import org.springframework.stereotype.Service;
-
-import wooteco.subway.dao.StationDao;
+import wooteco.subway.dao.StationDaoImpl;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
@@ -13,26 +11,26 @@ import wooteco.subway.dto.StationResponse;
 @Service
 public class StationService {
 
-    private final StationDao stationDao;
+    private final StationDaoImpl stationDaoImpl;
 
-    public StationService(StationDao stationDao) {
-        this.stationDao = stationDao;
+    public StationService(StationDaoImpl stationDaoImpl) {
+        this.stationDaoImpl = stationDaoImpl;
     }
 
     public StationResponse create(StationRequest stationRequest) {
         String name = stationRequest.getName();
-        Station station = stationDao.save(name);
+        Station station = stationDaoImpl.save(name);
         return new StationResponse(station);
     }
 
     public List<StationResponse> show() {
-        List<Station> stations = stationDao.findAll();
+        List<Station> stations = stationDaoImpl.findAll();
         return stations.stream()
-                .map(StationResponse::new)
-                .collect(Collectors.toList());
+            .map(StationResponse::new)
+            .collect(Collectors.toList());
     }
 
     public void delete(Long id) {
-        stationDao.deleteById(id);
+        stationDaoImpl.deleteById(id);
     }
 }
