+++ src/test/java/wooteco/subway/dao/StationDaoTest.java
@@ -1,16 +1,15 @@
 package wooteco.subway.dao;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.TestConstructor;
-
 import wooteco.subway.domain.Station;
 
 @SpringBootTest
@@ -50,8 +49,8 @@ public class StationDaoTest {
         String expected = "선릉역";
 
         assertThatThrownBy(() -> stationDao.save(expected))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이미 존재하는 지하철역 이름입니다.");
+            .isInstanceOf(IllegalArgumentException.class)
+            .hasMessage("이미 존재하는 지하철역 이름입니다.");
     }
 
     @Test
@@ -69,7 +68,7 @@ public class StationDaoTest {
     void deleteById() {
         stationDao.deleteById(1L);
 
-        assertThat(stationDao.findAll()).hasSize(0);
+        assertThat(stationDao.findAll()).isEmpty();
     }
 }
